refactor: Simplify controller.js and ui.js files
I made these changes because I realised I was getting sidetracked. I was trying to make everything modular and stick to the single responsibility principle, but it ended up making things more complicated than they needed to be.
The main point here is to learn about using async code, and I lost sight of that.
By simplifying the code here, it is now easier to follow what is going on and focus on using async code, which is the main goal of this project.