feat: format date and time using date-fns
This commit adds two simple date formatting functions to the utility file using the date-fns library to display the date and time in a more aesthetically pleasing format.

These functions are then imported into the api.js file to format the date, finally in ui.js a simple print out of the data has been used for testing purposes.