feat: Introduce delay to loading spinner
Reason:
Previously when the spinner functions were called and the user has a fast internet connection, the loading spinner would still display on the screen, however its so fast it appears like a glitch in the app.

To prevent the spinner from appearing as a brief glitch on fast connections, a 300ms delay has been added.
This ensures that the spinner only shows for slower network connections, providing a smoother UX.