fix: Ensure spinner covers entire paget
Problem:
During testing, I found users could still interact with the search form even when the spinner overlay was active.
This was not the intended behavior, as the spinner is meant to indicate that the system is processing and the form should be non-interactive.

Fix:
Adjusting the spinner overlay's z-index to a higher value ensures that it properly covers all the elements of the search, rendering the entire form non-interactive during loading.