style(ui): switch to flex-col at sm screen size
This commit adds styling to the forecast days elements and switches the to flex-column layout to avoid overflow issues on small screens.

Additionally at the smaller size each day increases to w-48 to fill some of the extra white space that is created around the elements when the flex-col style takes effect.